+++ ord_schema/scripts/check_pb.py
@@ -25,7 +25,7 @@ import pprint
 
 from google.protobuf import text_format
 
-from ord_schema import message_helpers
+from ord_schema import datasets
 
 
 def parse_args(argv: list[str] | None = None) -> argparse.Namespace:
@@ -38,7 +38,7 @@ def parse_args(argv: list[str] | None = None) -> argparse.Namespace:
 
 def main(args: argparse.Namespace) -> None:
     """Verifies that the pb and pbtxt Datasets are identical, raising on any difference."""
-    dataset = message_helpers.load_dataset(args.pb)
+    dataset = datasets.load_dataset(args.pb)
     pb_data = text_format.MessageToString(dataset)
     with pathlib.Path(args.pbtxt).open() as f:
         pbtxt_data = f.read()
